Start app refactoring
