--- crawlclima/config/cemaden.py
@@ -1,7 +1,0 @@
-"""
-Dados para acesso aos Webservices do CEMADEN
-"""
-
-chave = 'bc10602ea62759fab1578f8eb1ff6f7abbf8678d'
-url_rede = 'http://150.163.255.246:18383/dados_rede'
-url_pcd = 'http://150.163.255.246:18383/dados_pcd'
