+++ pyproject.toml
@@ -45,6 +45,9 @@ pre-commit = "^2.20.0"
 isort = "^5.10.1"
 flake8 = "^5.0.4"
 
+[tool.poetry.group.dev.dependencies]
+responses = "^0.22.0"
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
