--- utilities/env_sample
@@ -1,13 +1,0 @@
-PSQL_HOST=localhost
-PSQL_DB=dengue
-PSQL_USER=test_user
-PSQL_PASSWORD=xzh231
-PSQL_PORT=25432
-
-PSQL_HOST=localhost
-PSQL_DB=dengue
-PSQL_USER=test_user
-PSQL_PASSWORD=xzh231
-PSQL_PORT=5432=25432
-
-GEOJSON_PATH="GEOJSON_PATH"
